Buffs Rage Mage.
-- Rage Mage honestly feels very weak by the time you hit the orc caves.
I think there are a couple of reasons for this:

1. Almost half of their spells are out of combat utility spells. Despite having
magic fueled by battle, and not needing to carry books (so they can rely on
items for utility). Furthermore, most of their spells that aren't utility are
only effective against spellcasters. This plus their low max blow count (only 3!)
means they're reduced to just shouting at non-spellcasters. Boring, and not
particularly effective. Especially one on one.

2. Awesome Blow is the primary source of damage dealing for a Rage Mage
by the time they hit the orc caves, but it's way too unreliable. It has a
chance to fail, it is *very* expensive, and you can even miss with your attack!
Plus it only does a ton of damage if you position enemies against a wall.

So, we are buffing Awesome Blows, and replacing the utility spells in the hopes
of making this class more powerful and fun . We double the duration of all the
buffs to account for each one taking twice as long to cast. Awesome Blow
is given a very large buff to accuracy is given to it, so that it
hits reliably.

A subsequent commit will be replacing the utility spells with more combat-useful
spells. These spells will be geared towards making the Rage Mage more effective
in melee and when outnumbered. Finally, a future commit will also increase their
blow count, so that they can actually do some damage.